% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{curedata}
\alias{curedata}
\title{Simulated data for testing mixture cure models}
\format{
\code{t} Survival times, right-censored at 10 years.

\code{status}. 1 for an observed death and 0 for censoring.

\code{x}. A numeric variable with value of 0 for 100 individuals, and 1 for the other 100.
}
\source{
Simulated.
}
\usage{
curedata
}
\description{
Survival times of 200 fake people.  The cure probability is 0.5
for \code{x=0}, and the log odds ratio for cure is 0.5, so that the
cure probability is about 0.62 for \code{x=1}.  The uncured population
have survival times distributed as a Weibull with shape 1.5 and
scale 1.2.
}
\keyword{datasets}
