% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{get_draws}
\alias{get_draws}
\title{Posterior draws from a survextrap model}
\usage{
get_draws(x)
}
\arguments{
\item{x}{A fitted model object as returned by \code{\link{survextrap}}}
}
\description{
Return the matrix of draws from the posterior distribution of parameters in a survextrap model,
with all chains collapsed.
}
