% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{hazard_ratio}
\alias{hazard_ratio}
\title{Hazard ratio against time in a survextrap model}
\usage{
hazard_ratio(
  x,
  newdata = NULL,
  times = NULL,
  tmax = NULL,
  niter = NULL,
  sample = FALSE
)
}
\arguments{
\item{x}{A fitted model object as returned by \code{\link{survextrap}}}

\item{newdata}{A data frame with two rows.  The hazard ratio will be defined as hazard(first row)
divided by hazard(second row).}

\item{times}{Vector of times at which to compute the estimates.}

\item{tmax}{Maximum time at which to compute the estimates.  If
\code{times} is supplied, then this is ignored.  If \code{times} is not
supplied, then \code{times} is set to a set of 100 equally spaced
time points from 0 to \code{tmax}.  If both \code{tmax} and \code{times}
are not supplied, then \code{tmax} is set to the maximum follow up time
in the data.}

\item{niter}{Number of MCMC iterations to use to compute credible
intervals.  Set to a low value to make this function quicker, at the cost of
some approximation error (which may not be important for plotting).}

\item{sample}{If \code{TRUE} then the MCMC samples are returned instead
of being summarised as a median and 95\% credible intervals.}
}
\description{
Intended for use with non-proportional hazards models (\code{survextrap(...,nonprop=TRUE)}).
}
