% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{hrtime}
\alias{hrtime}
\title{Hazard ratio between high and low values of the hazard over time}
\usage{
hrtime(x, newdata = NULL, niter = NULL, hq = c(0.1, 0.9))
}
\arguments{
\item{x}{A fitted model object as returned by \code{\link{survextrap}}}

\item{newdata}{Data frame of covariate values to compute the output for.
If there are covariates in the model and this is not supplied,
the following default is used:

(a) if the only covariate is one factor variable, then the output is computed
for each level of this factor.

(b) if there are multiple covariates, or any numeric covariates, then the output
is computed at the mean of each numeric covariate in the original data, and at the
baseline level of each factor covariate.}

\item{niter}{Number of MCMC iterations to use to compute credible
intervals.  Set to a low value to make this function quicker, at the cost of
some approximation error (which may not be important for plotting).}

\item{hq}{Quantiles which define the "low" and "high" values to compute the hazard ratio between.
By default, this is \code{c(0.1, 0.9)}, so that the 10\\% and 90\\% quantiles are used
respectively.}
}
\value{
A summary of the posterior distribution of this hazard ratio from the fitted model,
as a data frame with one row per covariate value requested in \code{newdata}, and one column for each
of three posterior quantiles.
}
\description{
This is intended as an intuitive single-number measure of how much a hazard function changes over time.
The hazard is computed on an equally-spaced fine grid between the boundary knots.
The ratio between a "high" and "low" one of these hazard values is computed.
}
\details{
For example, if the hazard is constant over time, then this hazard ratio will be 1.
}
