% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspline.R
\name{mspline_args}
\alias{mspline_args}
\title{Documentation for common M-spline arguments}
\arguments{
\item{iknots}{Vector of internal knot locations. If not supplied, \code{df} has to be specified, in which case
the default is \code{df - degree - 1} equally spaced knots between the boundary knots.}

\item{bknots}{Vector with two elements, giving the boundary knot locations}

\item{degree}{Spline polynomial degree (defaults to 3)}

\item{df}{Desired number of basis terms, or "degrees of freedom" in the spline.
If \code{iknots} is not supplied, the number of internal knots is then chosen to satisfy this.}
}
\description{
Documentation for common M-spline arguments
}
