% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspline_dfns.R
\name{mspline_basis}
\alias{mspline_basis}
\title{Evaluate an M-spline basis matrix at the specified times.}
\usage{
mspline_basis(times, iknots, bknots, degree = 3, integrate = FALSE)
}
\arguments{
\item{times}{A numeric vector of times at which to evaluate the basis.}

\item{iknots}{Internal knots}

\item{bknots}{Boundary knots}

\item{degree}{Spline degree}

\item{integrate}{If \code{TRUE}, then the integrated M-spline (I-spline) basis is returned.}
}
\value{
A two-dimensional array.  Rows are the times, and columns are the basis terms.
}
\description{
Extrapolation beyond the boundary knots is done by assuming that each basis term
is constant beyond the boundary at its value at the boundary.
This gives a continuous but non-smooth function.   Each basis term is assumed to be
zero at times less than zero, since these models are used for hazard functions
in survival data.
}
