% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspline_plots.R
\name{mspline_priorpred}
\alias{mspline_priorpred}
\alias{plot_mspline_priorpred}
\title{Generate and/or plot a sample from the prior distribution of M-spline hazard curves}
\usage{
mspline_priorpred(
  iknots = NULL,
  bknots = c(0, 10),
  df = 10,
  degree = 3,
  coefs_mean = NULL,
  prior_smooth = p_gamma(2, 1),
  prior_loghaz = NULL,
  prior_loghr = NULL,
  x = NULL,
  X = NULL,
  X0 = NULL,
  nonprop = FALSE,
  prior_sdnp = p_gamma(2, 1),
  tmin = 0,
  tmax = 10,
  nsim = 10
)

plot_mspline_priorpred(
  iknots = NULL,
  bknots = c(0, 10),
  df = 10,
  degree = 3,
  coefs_mean = NULL,
  prior_smooth = p_gamma(2, 1),
  prior_loghaz = p_normal(0, 20),
  prior_loghr = NULL,
  x = NULL,
  X = NULL,
  nonprop = FALSE,
  prior_sdnp = p_gamma(2, 1),
  tmin = 0,
  tmax = 10,
  nsim = 10
)
}
\arguments{
\item{iknots}{Vector of internal knot locations. If not supplied, \code{df} has to be specified, in which case
the default is \code{df - degree - 1} equally spaced knots between the boundary knots.}

\item{bknots}{Vector with two elements, giving the boundary knot locations}

\item{df}{Desired number of basis terms, or "degrees of freedom" in the spline.
If \code{iknots} is not supplied, the number of internal knots is then chosen to satisfy this.}

\item{degree}{Spline polynomial degree (defaults to 3)}

\item{coefs_mean}{Spline basis coefficients that define the prior mean for the hazard function. By
default, these are set to values that define a constant hazard function.  They are normalised to
sum to 1 internally (if they do not already).}

\item{prior_smooth}{Gamma prior for the smoothing standard deviation, in
models where this is estimated.  This should be a call to \code{p_gamma()}.  The
default is \code{p_gamma(2,1)}.}

\item{prior_loghaz}{Prior for the baseline log rate parameter (\code{log(eta)}), after centering around the log
crude event rate in the data.
This should be a call to a prior constructor function, such as
\code{p_normal(0,1)} or \code{p_t(0,2,2)}.   Supported prior distribution families
are normal (parameters mean and SD) and t distributions (parameters
location, scale and degrees of freedom).  The default is a normal distribution with
mean 0 and standard deviation 20.

"Baseline" is defined
by the continuous covariates taking a value of zero and factor covariates taking their
reference level.  To use a different baseline, the data should be transformed
appropriately beforehand, so that a value of zero has a different meaning.

For continuous covariates, it helps for both computation and interpretation to define the value of zero to
denote a typical value in the data, e.g. the mean.}

\item{prior_loghr}{Priors for log hazard ratios.  This should be a call to
\code{p_normal()} or \code{p_t()}.  A list of calls can also be provided, to give
different priors to different coefficients, where the name
of each list component matches the name of the coefficient, e.g.
\code{list("age45-59" = p_normal(0,1), "age60+" = p_t(0,2,3))}

The default is \code{p_normal(0,2.5)} for all coefficients.}

\item{x}{Only required for models with covariates.  A list with components \code{ncovs} (number of covariates)
and \code{xnames} (names of these covariates).}

\item{X}{Either a list, vector or one-row matrix with one covariate value.
Factors must be supplied as multiple numeric (0/1) contrasts. (TODO doc on how to build
the names - vignette)}

\item{X0}{An alternative covariate value (optional).  The hazard ratio between the hazards at X
and X0 will also be returned.}

\item{nonprop}{Is the model a non-proportional hazards model (\code{TRUE} or \code{FALSE}).}

\item{prior_sdnp}{Prior for the standard deviation parameters that smooth the non-proportionality
effects over time in non-proportional hazards models.  This should be a call to \code{p_gamma()}
or a list of calls to \code{p_gamma()} with one component per covariate, as in \code{prior_loghr}.}

\item{tmin}{Minimum plotting time.  Defaults to lower boundary knot.}

\item{tmax}{Maximum plotting time.  Defaults to upper boundary knot.}

\item{nsim}{Number of simulations to draw}
}
\value{
\code{mspline_priorpred} returns a data frame of the samples, and
\code{plot_mspline_priorpred} generates a plot.
}
\description{
Generates and/or plots the hazard curves (as functions of time) implied by
a prior mean for the spline coefficients (a constant hazard by default)
and particular priors for the baseline log hazard and smoothness variance
}
