% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspline_plots.R
\name{mspline_uniform_weights}
\alias{mspline_uniform_weights}
\title{Estimate M-spline basis weights which give a constant hazard}
\usage{
mspline_uniform_weights(knots, bknots, times = NULL, degree = 3)
}
\arguments{
\item{knots}{TODO}

\item{bknots}{TODO}

\item{times}{TODO}

\item{degree}{TODO}
}
\description{
Estimate M-spline basis weights which give a constant hazard
}
