% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{p_hr}
\alias{p_hr}
\title{Derive a normal prior for the log hazard ratio parameter based on a guess at the hazard ratio}
\usage{
p_hr(median, upper)
}
\arguments{
\item{median}{Best guess (prior median) for a typical hazard ratio}

\item{upper}{Upper limit of 95\% credible interval for hazard ratio}
}
\value{
A normal prior in the format returned by \code{\link{p_normal}}, which can
be passed directly to the \code{prior_loghr} argument in \code{\link{survextrap}}.
}
\description{
Derive a normal prior for the log hazard ratio parameter based on a guess at the hazard ratio.
This assumes that the log upper limit is 2 standard deviations away from the log median.
}
