% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.survextrap}
\alias{plot.survextrap}
\title{Plot method for survextrap model objects}
\usage{
\method{plot}{survextrap}(x, type = "hazsurv", ...)
}
\arguments{
\item{x}{Fitted model object from \code{\link{survextrap}}.}

\item{type}{\code{"survival"} for a plot of the survival function, \code{"hazard"} for the hazard function, against time.}

\item{...}{Additional arguments, passed on to \code{plot_hazard} and \code{plot_survival}.}
}
\description{
Plot method for survextrap model objects
}
