% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_hazard_ratio}
\alias{plot_hazard_ratio}
\title{Plot hazard ratio against time from a survextrap model}
\usage{
plot_hazard_ratio(
  x,
  newdata = NULL,
  times = NULL,
  tmax = NULL,
  niter = NULL,
  ci = TRUE,
  xlab = "Time",
  ylab = "Hazard ratio",
  line_size = 1.5,
  ci_alpha = 0.2
)
}
\arguments{
\item{x}{A fitted model object as returned by \code{\link{survextrap}}}

\item{newdata}{A data frame with two rows.  The hazard ratio will be defined as hazard(first row)
divided by hazard(second row).}

\item{times}{Vector of times at which to compute the estimates.}

\item{tmax}{Maximum time at which to compute the estimates.  If
\code{times} is supplied, then this is ignored.  If \code{times} is not
supplied, then \code{times} is set to a set of 100 equally spaced
time points from 0 to \code{tmax}.  If both \code{tmax} and \code{times}
are not supplied, then \code{tmax} is set to the maximum follow up time
in the data.}

\item{niter}{Number of MCMC iterations to use to compute credible
intervals.  Set to a low value to make this function quicker, at the cost of
some approximation error (which may not be important for plotting).}

\item{ci}{If \code{TRUE} then credible intervals are drawn.  Defaults to
drawing the intervals if the plot shows the curve for only one covariate value.}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}

\item{line_size}{Passed to \code{\link{geom_line}}.}

\item{ci_alpha}{Transparency for the credible interval ribbons}
}
\description{
Intended for use with non-proportional hazards models (\code{survextrap(...,nonprop=TRUE)}).
}
