% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspline_plots.R
\name{plot_mspline_priorpred}
\alias{plot_mspline_priorpred}
\title{Generate a sample from the prior distribution of M-spline hazard curves implied by
a particular mean and variance for the baseline curve and scale parameter.}
\usage{
plot_mspline_priorpred(
  knots = NULL,
  bknots = c(0, 10),
  df = 10,
  degree = 3,
  prior_mean,
  prior_sd = 1,
  scale = 1,
  scale_sd = 1,
  tmin = 0,
  tmax = 10,
  nsim = 10
)
}
\arguments{
\item{knots}{Internal knots.  If not supplied, df has to be specified.  Then defaults
to equally spaced knots between the boundary knots.}

\item{bknots}{Boundary knots.   Defaults to \code{c(0,10)}.}

\item{df}{df TODO}

\item{degree}{Spline polynomial degree.}

\item{prior_mean}{Prior mean TODO document properly}

\item{prior_sd}{Prior SD TODO}

\item{scale}{Prior mean for scale parameter TODO}

\item{scale_sd}{Prior SD for scale parameter TODO}

\item{tmin}{Minimum plotting time.  Defaults to lower boundary knot}

\item{tmax}{Maximum plotting time.  Defaults to upper boundary knot}

\item{nsim}{Number of simulations to draw}
}
\description{
See the methods vignette for full info for now
}
