% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.survextrap}
\alias{print.survextrap}
\title{Print a fitted survextrap model}
\usage{
\method{print}{survextrap}(x, ...)
}
\arguments{
\item{x}{A fitted model object as returned by \code{\link{survextrap}}}

\item{...}{Other arguments (currently unused).}
}
\description{
Print a fitted survextrap model
}
