% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_priors}
\alias{print_priors}
\title{Print the priors used in a fitted survextrap model}
\usage{
print_priors(x)
}
\arguments{
\item{x}{A fitted model object as returned by \code{\link{survextrap}}}
}
\value{
No explicit value is returned.
}
\description{
Print the priors used in a fitted survextrap model
}
