% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{prior_haz_const}
\alias{prior_haz_const}
\title{Summarises the prior for the constant hazard implied by a particular
prior on the hazard scale parameter and spline specification.}
\usage{
prior_haz_const(
  mspline,
  prior_loghaz = p_normal(0, 20),
  nsim = 10000,
  quantiles = c(0.025, 0.5, 0.975)
)
}
\arguments{
\item{mspline}{A list of control parameters defining the spline model.

\code{iknots}: Internal knots.  If this is not supplied, then the number of
knots is taken from \code{df}, and their location is taken from equally-spaced
quantiles of the observed event times (concatenated with the distinct
follow-up times in the external data).

\code{bknots}: Boundary knots.  If this is not supplied, the boundary knots are
set to zero and the maximum event time (including the follow-up times in
the external data).

\code{df}: Degrees of freedom, i.e. the number of parameters (or basis terms)
that define the hazard as a spline function of time.  Defaults to 10.  Note
this does not necessarily overfit because the function is smoothed through
the prior.

\code{degree}: Polynomial degree used for the basis function. The default is 3, giving a cubic.}

\item{prior_loghaz}{Prior for the baseline log hazard scale parameter (\code{log(eta)}).
This should be a call to a prior constructor function, such as
\code{p_normal(0,1)} or \code{p_t(0,2,2)}.   Supported prior distribution families
are normal (parameters mean and SD) and t distributions (parameters
location, scale and degrees of freedom).  The default is a normal distribution with
mean 0 and standard deviation 20.

Note that \code{eta} is not in itself a hazard, but it is proportional to the hazard (see the vignette for the
full model specification).

"Baseline" is defined
by the continuous covariates taking a value of zero and factor covariates taking their
reference level.  To use a different baseline, the data should be transformed
appropriately beforehand, so that a value of zero has a different meaning.

For continuous covariates, it helps for both computation and interpretation to define the value of zero to
denote a typical value in the data, e.g. the mean.}

\item{nsim}{Number of simulations to draw}

\item{quantiles}{Quantiles used to summarise the prior predictive distribution
of the standard deviation.}
}
\description{
Summarises the prior for the constant hazard implied by a particular
prior on the hazard scale parameter and spline specification.
}
