% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_pred.R
\name{prior_pred}
\alias{prior_pred}
\title{Simulate a dataset from the prior predictive distribution of survival times
in an M-spline survival model.}
\usage{
prior_pred(
  n,
  censtime = Inf,
  alpha_mean = 0,
  alpha_sd = 20,
  lcoef_mean = NULL,
  lcoef_sd = 1,
  iknots,
  bknots,
  degree = 3
)
}
\arguments{
\item{n}{Sample size of the simulated dataset.  Each observation in the dataset is
generated from a model with the same parameters.  These parameters are generated from
a single simulation from the prior distribution.}

\item{censtime}{Right-censoring time to impose on the simulated event times.}

\item{alpha_mean}{Prior mean of the normal distribution for the baseline log hazard.}

\item{alpha_sd}{Prior standard deviation of the normal distribution for the baseline log hazard.}

\item{lcoef_mean}{Prior mean vector of the the logistic distribution for the spline coefficients on the multinomial logit scale.}

\item{lcoef_sd}{Prior standard deviation of the logistic distribution for the spline coefficients on the multinomial logit scale.}

\item{iknots}{Internal knots of the M-spline.}

\item{bknots}{Boundary knots of the M-spline.}

\item{degree}{M-spline polynomial degree.}
}
\value{
A data frame with columns \code{time} (simulated time) and \code{event} (indicator for whether
the time is an event time, as opposed to a right-censoring time).  The prior parameters are returned
in the \code{prior} attribute as a list with components \code{alpha} (baseline log hazard) and
\code{coefs} (spline coefficients).
}
\description{
Simulate a dataset from the prior predictive distribution of survival times
in an M-spline survival model.
}
