% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{prior_sample}
\alias{prior_sample}
\title{Draw a sample from the prior distribution of the parameters
governing the hazard in a survextrap model for given covariates.}
\usage{
prior_sample(
  mspline,
  coefs_mean = NULL,
  prior_smooth = p_gamma(2, 1),
  prior_loghaz,
  prior_loghr = NULL,
  X = NULL,
  X0 = NULL,
  prior_sdnp = NULL,
  nsim = 100
)
}
\arguments{
\item{mspline}{A list of control parameters defining the spline model.

\code{iknots}: Internal knots.  If this is not supplied, then the number of
knots is taken from \code{df}, and their location is taken from equally-spaced
quantiles of the observed event times (concatenated with the distinct
follow-up times in the external data).

\code{bknots}: Boundary knots.  If this is not supplied, the boundary knots are
set to zero and the maximum event time (including the follow-up times in
the external data).

\code{df}: Degrees of freedom, i.e. the number of parameters (or basis terms)
that define the hazard as a spline function of time.  Defaults to 10.  Note
this does not necessarily overfit because the function is smoothed through
the prior.

\code{degree}: Polynomial degree used for the basis function. The default is 3, giving a cubic.}

\item{coefs_mean}{Spline basis coefficients that define the prior mean for the hazard function. By
default, these are set to values that define a constant hazard function.  They are normalised to
sum to 1 internally (if they do not already).}

\item{prior_smooth}{Gamma prior for the smoothing standard deviation, in
models where this is estimated.  This should be a call to \code{p_gamma()}.  The
default is \code{p_gamma(2,1)}.}

\item{prior_loghaz}{Prior for the baseline log hazard scale parameter (\code{log(eta)}).
This should be a call to a prior constructor function, such as
\code{p_normal(0,1)} or \code{p_t(0,2,2)}.   Supported prior distribution families
are normal (parameters mean and SD) and t distributions (parameters
location, scale and degrees of freedom).  The default is a normal distribution with
mean 0 and standard deviation 20.

Note that \code{eta} is not in itself a hazard, but it is proportional to the hazard (see the vignette for the
full model specification).

"Baseline" is defined
by the continuous covariates taking a value of zero and factor covariates taking their
reference level.  To use a different baseline, the data should be transformed
appropriately beforehand, so that a value of zero has a different meaning.

For continuous covariates, it helps for both computation and interpretation to define the value of zero to
denote a typical value in the data, e.g. the mean.}

\item{prior_loghr}{Priors for log hazard ratios.  This should be a call to
\code{p_normal()} or \code{p_t()}.  A list of calls can also be provided, to give
different priors to different coefficients, where the name
of each list component matches the name of the coefficient, e.g.
\code{list("age45-59" = p_normal(0,1), "age60+" = p_t(0,2,3))}

The default is \code{p_normal(0,2.5)} for all coefficients.}

\item{X}{Either a list, vector or one-row matrix with one covariate value.
Factors must be supplied as multiple numeric (0/1) contrasts. (TODO doc on how to
match then with prior_loghr and prior_sdnp, and naming)}

\item{X0}{An alternative covariate value (optional).  The hazard
ratio between the hazards at X and X0 will also be returned.}

\item{prior_sdnp}{Prior for the standard deviation parameters that smooth the non-proportionality
effects over time in non-proportional hazards models.  This should be a call to \code{p_gamma()}
or a list of calls to \code{p_gamma()} with one component per covariate, as in \code{prior_loghr}.}

\item{nsim}{Number of simulations to draw}
}
\value{
A list with components:

\code{loghaz}: Baseline log scale parameter (\code{log(eta)} in the notation of the manual). For models with covariates, this is at the covariate values supplied in \code{X}, or at zero if \code{X} is not supplied.

\code{haz}: Baseline scale parameter (\code{eta}).

\code{coefs}: Spline coefficients. For non-proportional hazards model with covariates, these are returned at the suppled value of \code{X}, or at values of zero if \code{X} is not supplied.

\code{beta}: Multinomial logit-transformed spline coefficients.

\code{smooth}: Smoothing standard deviation for spline coefficients.
}
\description{
This is used, for example, in \code{\link{mspline_priorpred}} to visualise the
prior distribution around hazard curves implied by a particular M-spline model
and parameter priors.
}
\details{
Cure and relative survival models are not currently handled.
}
