% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspline_plots.R, R/priors.R
\name{plot_prior_hazard}
\alias{plot_prior_hazard}
\alias{prior_sample_hazard}
\title{Generate and/or plot a sample from the prior distribution of M-spline hazard curves}
\usage{
plot_prior_hazard(
  knots = NULL,
  df = 10,
  degree = 3,
  coefs_mean = NULL,
  prior_hsd = p_gamma(2, 1),
  prior_hscale = p_normal(0, 20),
  prior_loghr = NULL,
  X = NULL,
  prior_hrsd = p_gamma(2, 1),
  tmin = 0,
  tmax = NULL,
  nsim = 10
)

prior_sample_hazard(
  knots = NULL,
  df = 10,
  degree = 3,
  bsmooth = TRUE,
  coefs_mean = NULL,
  prior_hsd = p_gamma(2, 1),
  prior_hscale = NULL,
  prior_loghr = NULL,
  X = NULL,
  X0 = NULL,
  prior_hrsd = NULL,
  tmin = 0,
  tmax = 10,
  nsim = 10
)
}
\arguments{
\item{knots}{Vector of knot locations. If not supplied, \code{df}
has to be specified, in which case the default is a set of
equally spaced knots between zero and the highest knot.  The
number of knots (excluding zero) is \code{df - degree + 1} if
\code{bsmooth} is \code{TRUE}, or \code{df - degree - 1}
otherwise.}

\item{df}{Desired number of basis terms, or "degrees of freedom"
in the spline.  If \code{knots} is not supplied, the number of
knots is then chosen to satisfy this.}

\item{degree}{Spline polynomial degree.  Can only be changed from
the default of 3 if \code{bsmooth} is \code{FALSE}.}

\item{coefs_mean}{Spline basis coefficients that define the prior
mean for the hazard function. By default, these are set to values
that define a constant hazard function (see
\code{\link{mspline_constant_coefs}}).  They are normalised to
sum to 1 internally (if they do not already).}

\item{prior_hsd}{Gamma prior for the standard deviation that
controls the variability over time (or smoothness) of the hazard
function.  This should be a call to \code{p_gamma()}.  The default is
\code{p_gamma(2,1)}.  See \code{\link{prior_haz_sd}} for a way to
calibrate this to represent a meaningful belief.}

\item{prior_hscale}{Prior for the baseline log hazard scale parameter (\code{alpha} or \code{log(eta)}).
This should be a call to a prior constructor function, such as
\code{p_normal(0,1)} or \code{p_t(0,2,2)}.   Supported prior distribution families
are normal (parameters mean and SD) and t distributions (parameters
location, scale and degrees of freedom).  The default is a normal distribution with
mean 0 and standard deviation 20.

Note that \code{eta} is not in itself a hazard, but it is proportional to the hazard (see the vignette for the
full model specification).

"Baseline" is defined by the continuous covariates taking a value
of zero and factor covariates taking their reference level.  To
use a different baseline, the data should be transformed
appropriately beforehand, so that a value of zero has a different
meaning.  For continuous covariates, it helps for both
computation and interpretation to define the value of zero to
denote a typical value in the data, e.g. the mean.}

\item{prior_loghr}{Priors for log hazard ratios.  This should be a call to
\code{p_normal()} or \code{p_t()}.  A list of calls can also be provided, to give
different priors to different coefficients, where the name
of each list component matches the name of the coefficient, e.g.
\code{list("age45-59" = p_normal(0,1), "age60+" = p_t(0,2,3))}

The default is \code{p_normal(0,2.5)} for all coefficients.}

\item{X}{Either a list, vector or one-row matrix, with one value
for each covariate in the model.  Defaults to zero.  Factors
must be supplied as multiple numeric (0/1) contrasts.
If there are multiple covariates and a single prior given in
\code{prior_loghr}, then this prior will be used for all the covariates.
If a list of priors is provided, then both this list and \code{X}
must be named with the names of the covariates.}

\item{prior_hrsd}{Prior for the standard deviation parameters that
smooth the non-proportionality effects over time in
non-proportional hazards models.  This should be a call to
\code{p_gamma()} or a list of calls to \code{p_gamma()} with one component
per covariate, as in \code{prior_loghr}.  See
\code{\link{prior_hr_sd}} for a way to calibrate this to
represent a meaningful belief.}

\item{tmin}{Minimum plotting time.  Defaults to zero.}

\item{tmax}{Maximum plotting time.  Defaults to the highest knot.}

\item{nsim}{Number of simulations to draw}

\item{bsmooth}{If \code{TRUE} then the function is constrained to
also have zero derivative and second derivative at the boundary.}

\item{X0}{An alternative set of "reference" covariate values
(optional).  The hazard ratio between the hazards at \code{X}
and \code{X0} will also be returned.}
}
\value{
\code{prior_sample_hazard} returns a data frame of the
samples, and \code{plot_prior_hazard} generates a plot.  No
customisation options are provided for the plot function, which
is just intended as a quick check.
}
\description{
Generates and/or plots the hazard curves (as functions of time)
implied by a prior mean for the spline coefficients (a constant
hazard by default) and particular priors for the baseline log
hazard and smoothness standard deviation.
}
