% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{rmst_survextrap}
\alias{rmst_survextrap}
\title{Restricted mean survival time}
\usage{
rmst_survextrap(x, t, newdata = NULL, niter = NULL)
}
\arguments{
\item{x}{A fitted model object as returned by \code{\link{survextrap}}}

\item{t}{Vector of times.  The restricted mean survival time up to each one of these times
will be computed.}

\item{newdata}{Data frame of covariate values to compute the output for.
If there are covariates in the model and this is not supplied,
the following default is used:

(a) if the only covariate is one factor variable, then the output is computed
for each level of this factor.

(b) if there are multiple covariates, or any numeric covariates, then the output
is computed at the mean of each numeric covariate in the original data, and at the
baseline level of each factor covariate.}

\item{niter}{Number of MCMC iterations to use to compute credible
intervals.  Set to a low value to make this function quicker, at the cost of
some approximation error (which may not be important for plotting).}
}
\description{
Restricted mean survival time
}
