% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.survextrap}
\alias{summary.survextrap}
\title{Posterior summary statistics for parameters of survextrap models}
\usage{
\method{summary}{survextrap}(object, ...)
}
\arguments{
\item{object}{A fitted model object as returned by \code{\link{survextrap}}}

\item{...}{Other argument (currently unused).

Document meanings of parameters

\code{alpha}: Average baseline log hazard.  If there are covariates, this describes the
average hazard with continuous
covariates set to zero, and factor covariates set
to their baseline levels.

\code{coefs}: Coefficients of the M-spline basis terms.

\code{loghr}: Log hazard ratios for each covariate in the model. For cure
models, this refers to covariates on survival for uncured people.

\code{hr}: Hazard ratios (the exponentials of \code{loghr}).

\code{pcure}: Probability of cure (for cure models only).   If there are covariates
on cure, this parameter describes the probability of cure with continuous
covariates set to zero, and factor covariates set
to their baseline levels.

\code{logor_cure}: Log odds ratio of cure for each covariate on cure.

\code{or_cure}: Odds ratios of cure (the exponentials of \code{logor_cure}).}
}
\description{
The intervals are 95\\% credible intervals.
}
\details{
Suggestions for what else to add to this are welcome.  Convergence diagnostics?
Mode for optim method.  Mean?
At least document how to get other arbitrary posterior summaries using
the posterior package.  Could even allow as an argument to this.
}
