% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survextrap-package.R
\docType{package}
\name{survextrap-package}
\alias{survextrap-package}
\title{The 'survextrap' package.}
\description{
For an introduction to and overview of the \code{survextrap} package, and full documentation, see

\url{https://chjackson.github.io/survextrap/}.

For details of the methods, see the paper by Jackson (2023).
}
\references{
Jackson, C. (2023) \code{survextrap}: a package for flexible and transparent
survival extrapolation.  arXiV preprint, \url{https://arxiv.org/abs/2306.03957}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chjackson/survextrap}
  \item \url{https://chjackson.github.io/survextrap/}
  \item Report bugs at \url{https://github.com/chjackson/survextrap/issues}
}

}
\author{
\strong{Maintainer}: Christopher Jackson \email{chris.jackson@mrc-bsu.cam.ac.uk} (\href{https://orcid.org/0000-0002-6656-8913}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Iain Timmins (\href{https://orcid.org/0000-0002-8761-582X}{ORCID}) [contributor]
}

}
