% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{survival}
\alias{survival}
\title{Estimates of survival from a survextrap model}
\usage{
survival(
  x,
  newdata = NULL,
  times = NULL,
  tmax = NULL,
  niter = NULL,
  sample = FALSE
)
}
\arguments{
\item{x}{A fitted model object as returned by \code{\link{survextrap}}}

\item{newdata}{Data frame of covariate values to compute the output for.
If there are covariates in the model and this is not supplied,
the following default is used:

(a) if the only covariate is one factor variable, then the output is computed
for each level of this factor.

(b) if there are multiple covariates, or any numeric covariates, then the output
is computed at the mean of each numeric covariate in the original data, and at the
baseline level of each factor covariate.}

\item{times}{Vector of times at which to compute the estimates.}

\item{tmax}{Maximum time at which to compute the estimates.  If
\code{times} is supplied, then this is ignored.  If \code{times} is not
supplied, then \code{times} is set to a set of 100 equally spaced
time points from 0 to \code{tmax}.  If both \code{tmax} and \code{times}
are not supplied, then \code{tmax} is set to the maximum follow up time
in the data.}

\item{niter}{Number of MCMC iterations to use to compute credible
intervals.  Set to a low value to make this function quicker, at the cost of
some approximation error (which may not be important for plotting).}

\item{sample}{If \code{TRUE} then the MCMC samples are returned instead
of being summarised as a median and 95\% credible intervals.}
}
\description{
Estimates of survival from a survextrap model
}
