% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspline_dfns.R
\name{Survmspline}
\alias{Survmspline}
\alias{dsurvmspline}
\alias{psurvmspline}
\alias{qsurvmspline}
\alias{rsurvmspline}
\alias{hsurvmspline}
\alias{Hsurvmspline}
\alias{mean_survmspline}
\alias{rmst_survmspline}
\title{M-spline survival distribution}
\usage{
psurvmspline(
  q,
  alpha,
  coefs,
  knots,
  degree = 3,
  lower.tail = TRUE,
  log.p = FALSE
)

Hsurvmspline(x, alpha, coefs, knots, degree = 3, log = FALSE)

hsurvmspline(x, alpha, coefs, knots, degree = 3, log = FALSE)

dsurvmspline(x, alpha, coefs, knots, degree = 3, log = FALSE)

qsurvmspline(
  p,
  alpha,
  coefs,
  knots,
  degree = 3,
  lower.tail = TRUE,
  log.p = FALSE
)

rsurvmspline(n, alpha, coefs, knots, degree = 3)

rmst_survmspline(t, alpha, coefs, knots, degree = 3)

mean_survmspline(alpha, coefs, knots, degree = 3)
}
\arguments{
\item{alpha}{Log scale parameter.}

\item{coefs}{Spline basis coefficients. These should sum to 1,
otherwise they are normalised internally to sum to 1.}

\item{knots}{Locations of knots on the axis of time, supplied in
increasing order.  These include the two boundary knots.

In vectorised usage of these functions, the knots and degree must be
the same for all alternative times and parameter values.}

\item{degree}{Spline polynomial degree.}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P(X
\le x)}{P(X <= x)}, otherwise, \eqn{P(X > x)}{P(X > x)}.}

\item{x, q, t}{Vector of times.}

\item{log, log.p}{Return log density or probability.}

\item{p}{Vector of probabilities.}

\item{n}{Number of random numbers to simulate.}
}
\value{
\code{dsurvmspline} gives the density, \code{psurvmspline} gives the
distribution function, \code{hsurvmspline} gives the hazard and
\code{Hsurvmspline} gives the cumulative hazard.

\code{qsurvmspline} gives the quantile function, which is computed by
numerical inversion.

\code{rsurvmspline} generates random survival times by using
\code{qsurvmspline} on a sample of uniform random numbers.
}
\description{
Probability density, distribution, quantile, random generation, hazard,
cumulative hazard, mean and restricted mean functions for the
M-spline time-to-event model.
}
\references{
M-spline paper, rstanarm paper
}
\author{
Christopher Jackson \href{mailto:chris.jackson@mrc-bsu.cam.ac.uk}{chris.jackson@mrc-bsu.cam.ac.uk}
}
\keyword{distribution}
