% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{colons}
\alias{colons}
\title{Colon cancer survival data}
\format{
Documented in \code{\link[survival]{colon}}.
}
\source{
See \code{\link[survival]{colon}}.
}
\usage{
colons
}
\description{
Survival times of 191 patients from a trial of chemotherapy for
colon cancer.  This is the data provided in the \code{survival} package
as \code{colon}, but this is restricted to outcome of recurrence-free
survival, artificially censored at 3 years, and taking a 20\%
random sample.
}
\references{
See \code{\link[survival]{colon}}.
}
\keyword{datasets}
