% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspline_plots.R
\name{mspline_plotsetup}
\alias{mspline_plotsetup}
\title{Get times and basis for an illustration of an M-spline with given knots.}
\usage{
mspline_plotsetup(
  knots,
  bknots = c(0, 10),
  tmin = NULL,
  tmax = NULL,
  degree = 3,
  df = 10,
  extrap_model = "constant"
)
}
\arguments{
\item{knots}{Internal knots.  If not supplied, df has to be specified.  Then defaults
to equally spaced knots between the boundary knots.}

\item{bknots}{Boundary knots.   Defaults to \code{c(0,10)}.}

\item{tmin}{Minimum plotting time.  Defaults to lower boundary knot}

\item{tmax}{Maximum plotting time.  Defaults to upper boundary knot}

\item{degree}{Spline polynomial degree.}

\item{df}{df TODO}

\item{extrap_model}{If \code{"constant"} then the hazard after the final boundary
knot equals the hazard at this knot.  Otherwise the basis functions are simply
extended outside boundary knot.}
}
\description{
Get times and basis for an illustration of an M-spline with given knots.
}
