% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspline.R
\name{mspline_uniform_weights}
\alias{mspline_uniform_weights}
\title{Estimate M-spline basis weights which give a constant function.}
\usage{
mspline_uniform_weights(
  iknots,
  bknots,
  times = NULL,
  degree = 3,
  logit = FALSE
)
}
\arguments{
\item{iknots}{Internal knots.}

\item{bknots}{Boundary knots.}

\item{times}{Times to use to construct the numerical calculation.
By default, this is 20 equally-spaced times between the boundary knots.}

\item{degree}{Spline polynomial degree.}

\item{logit}{If \code{TRUE} then the multinomial logit transform of the coefficients
is returned.  This is a vector of length one less than the number of coefficients,
with the rth element defined by \eqn{log(coefs[r+1] / coefs[1])}.}
}
\description{
This works by choosing the basis coefficients that minimise the
variance between log hazard values at different time points.
It is used in \code{\link{survextrap}} to choose the default prior mean
for the hazard function.
}
