% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{p_meansurv}
\alias{p_meansurv}
\title{Derive a normal prior for the hazard scale parameter based on a guess at survival times}
\usage{
p_meansurv(median, upper, mspline = NULL)
}
\arguments{
\item{median}{Best guess (prior median) for a typical survival time}

\item{upper}{95\% credible limit for a survival time}

\item{mspline}{A list with components \code{iknots} (vector of internal knots), \code{bknots} (vector of boundary knots)
and \code{degree} (polynomial degree) defining an M-spline configuration.}
}
\description{
Derive a normal prior for the hazard scale parameter based on a guess at survival times
}
