% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspline_plots.R
\name{plot_mspline}
\alias{plot_mspline}
\title{Plot a M-spline function, showing how it is built up from its basis}
\usage{
plot_mspline(
  iknots = NULL,
  bknots = c(0, 10),
  df = 10,
  degree = 3,
  coefs = NULL,
  scale = 1,
  tmin = 0,
  tmax = 10
)
}
\arguments{
\item{iknots}{Vector of internal knot locations. If not supplied, \code{df} has to be specified, in which case
the default is \code{df - degree - 1} equally spaced knots between the boundary knots.}

\item{bknots}{Vector with two elements, giving the boundary knot locations}

\item{df}{Desired number of basis terms, or "degrees of freedom" in the spline.
If \code{iknots} is not supplied, the number of internal knots is then chosen to satisfy this.}

\item{degree}{Spline polynomial degree (defaults to 3)}

\item{coefs}{Coefficients of the spline basis terms.  These are normalised internally to sum to 1,
if they do not already sum to 1.}

\item{scale}{Scale parameter. After computing the standard M-spline function as a weighted sum of the basis
terms, the function is multiplied by \code{scale}.   The log of the scale is the parameter called
\code{alpha} in the results of a \code{survextrap} model, the intercept of the linear model on the log hazard.}

\item{tmin}{Minimum plotting time.  Defaults to lower boundary knot.}

\item{tmax}{Maximum plotting time.  Defaults to upper boundary knot.}
}
\description{
Plot a M-spline function, showing how it is built up from its basis
}
