% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspline_plots.R
\name{plot_mspline}
\alias{plot_mspline}
\title{Plot a M-spline function, showing how it is built up from its basis}
\usage{
plot_mspline(
  knots = NULL,
  bknots = c(0, 10),
  df = 10,
  degree = 3,
  p = NULL,
  scale = 1,
  tmin = 0,
  tmax = 10,
  plot = TRUE
)
}
\arguments{
\item{knots}{Internal knots}

\item{bknots}{Boundary knots}

\item{df}{TODO}

\item{degree}{Degree of polynomials defining the spline}

\item{p}{Weights TODO rename to coefs}

\item{scale}{Scale parameter}

\item{tmin}{Minimum time to plot}

\item{tmax}{Maximum time to plot to}

\item{plot}{Plot}
}
\description{
Plot a M-spline function, showing how it is built up from its basis
}
