% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.survextrap}
\alias{print.survextrap}
\title{Print a fitted survextrap model}
\usage{
\method{print}{survextrap}(x, ...)
}
\arguments{
\item{x}{A fitted model object as returned by \code{\link{survextrap}}}

\item{...}{Other arguments (currently unused).}
}
\value{
No value is returned.

This prints a summary of the data, a statement of the fitted
model, the prior distributions, and a table of summary statistics
of the posterior distributions of the model parameters.  For
descriptions of the parameters in this summary table, see
\code{\link{summary.survextrap}}.
}
\description{
Print a fitted survextrap model
}
