% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.survextrap}
\alias{summary.survextrap}
\title{Posterior summary statistics for parameters of survextrap models}
\usage{
\method{summary}{survextrap}(object, ...)
}
\arguments{
\item{object}{A fitted model object as returned by \code{\link{survextrap}}}

\item{...}{Other functions to compute a summary statistic from a vector
of posterior samples can be supplied here.   Many useful such functions are
provided with the \pkg{posterior} package. If the arguments are named,
then these are used to name the returned data frame.  See the examples below.}
}
\value{
A data frame (actually a tibble) of summary statistics for the model parameters is returned.

The parameters, as indicated in the \code{variable} column, are:

\code{alpha}: Average baseline log hazard.  If there are covariates, this describes the
average hazard with continuous
covariates set to zero, and factor covariates set
to their baseline levels.

\code{coefs}: Coefficients of the M-spline basis terms.  If a non-proportional hazards model
was fitted, these are with covariates set to zero or baseline levels.

\code{loghr}: Log hazard ratios for each covariate in the model. For cure
models, this refers to covariates on survival for uncured people.  For non-proportional hazards
models, these are the effects of covariates on the scale parameter \code{eta} and represent
"average" hazard ratios, from which departures are modelled (see the "methods"
vignette for a full description of this model).

\code{hr}: Hazard ratios (the exponentials of \code{loghr}).

\code{pcure}: Probability of cure (for cure models only).   If there are covariates
on cure, this parameter describes the probability of cure with continuous
covariates set to zero, and factor covariates set
to their baseline levels.

\code{logor_cure}: Log odds ratio of cure for each covariate on cure.

\code{or_cure}: Odds ratios of cure (the exponentials of \code{logor_cure}).

\code{nperr}: Standardised departures from proportional hazards in the non-proportional hazards model, defined as \eqn{b^{(np)}_{ks} / \sigma^{(np)}_s} (see the "methods" vignette for definitions of these).

\code{sd_np}: Smoothness standard deviations \eqn{\sigma^{(np)}_s} for the non-proportionality effects.
}
\description{
By default, a median and 95\% credible intervals are presented, alongside
the Rhat convergence diagnostic and the bulk effective sample size (see the \pkg{posterior} package).  For models
fitted by optimisation rather than MCMC, the posterior mode is also returned.
}
\details{
Any other posterior summary can be computed if the appropriate function to compute it
is supplied here.
}
\examples{
mod <- survextrap(Surv(years, status) ~ rx, data=colons, fit_method="opt")
summary(mod)
summary(mod, mean)
summary(mod, mean, ess_tail=posterior::ess_tail)

}
