% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.survextrap}
\alias{summary.survextrap}
\title{Posterior summary statistics for parameters of survextrap models}
\usage{
\method{summary}{survextrap}(object, ...)
}
\arguments{
\item{object}{A fitted model object as returned by \code{\link{survextrap}}}

\item{...}{Other argument (currently unused).

\code{alpha}

\code{coefs}

\code{loghr}

\code{hr}

\code{pcure}

\code{logor_cure}

\code{or_cure}}
}
\description{
The intervals are 95\\% credible intervals.
}
\details{
Suggestions for what else to add to this are welcome.  Convergence diagnostics?
}
