% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survextrap-package.R
\docType{package}
\name{survextrap-package}
\alias{survextrap-package}
\title{The 'survextrap' package.}
\description{
For an introduction to and overview of the \code{survextrap} package, and full documentation, see

\url{https://chjackson.github.io/survextrap/}
}
