% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survextrap-package.R
\docType{package}
\name{survextrap-package}
\alias{survextrap-package}
\title{The 'survextrap' package.}
\description{
For an introduction to and an overview of the \code{survextrap} package, see

\url{https://chjackson.github.io/survextrap/}
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.26.11. https://mc-stan.org
}
