% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survextrap-package.R
\docType{package}
\name{survextrap-package}
\alias{survextrap-package}
\title{The 'survextrap' package.}
\description{
For an introduction to and overview of the \code{survextrap} package, and full documentation, see

\url{https://chjackson.github.io/survextrap/}.

For details of the methods, see the paper by Jackson (2023).
}
\references{
Jackson, C. (2023) \code{survextrap}: a package for flexible and transparent
survival extrapolation.  arXiV preprint, \url{https://arxiv.org/abs/2306.03957}.
}
