% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survextrap.R
\name{survextrap}
\alias{survextrap}
\title{survextrap}
\usage{
survextrap(
  formula,
  data,
  external = NULL,
  smooth_sd = "bayes",
  prior_weights = NULL,
  cure = FALSE,
  cure_prior = c(1, 1),
  basehaz_ops = NULL,
  modelid = 1,
  algorithm = "mcmc",
  ...
)
}
\arguments{
\item{formula}{A survival formula in standard R formula syntax.  Covariates included
on the right hand side of the formula with be modelled with proportional hazards.}

\item{data}{Data frame containing variables in \code{formula}.}

\item{external}{External data as a data frame of aggregate survival counts with columns:

\code{start}: Start time

\code{start}: Follow-up time

\code{n}: Number of people alive at \code{start}

\code{r}: Number of those people who are still alive at \code{stop}

It is intended to add facilities to the package to produce this format of data from
other common forms of external data, e.g. registry or population data, or elicited judgements
about conditional survival}

\item{smooth_sd}{Smoothing parameter estimation.

\code{"bayes"}: the smoothing parameter is estimated by full Bayes.

\code{"eb"}: empirical Bayes is used.

Alternatively, if a number is supplied here, then the smoothing parameter is fixed to this number.}

\item{prior_weights}{Spline basis weights defining the prior mean for the hazard function. By
default these are set to values that define a constant hazard function.  (TODO: a tool for people
to convert other hazard forms into prior weights). ?? prior mean or beta?}

\item{cure}{If \code{TRUE} a mixture cure model is used.}

\item{cure_prior}{Beta shape parameters for the cure probability (vector of two)}

\item{basehaz_ops}{A list of control parameters

\code{knots}: Internal knots.  If this is not supplied, then \code{df} chosen quantiles of the uncensored survival times.

\code{bknots}: Boundary knots.  Zero and maximum time
?? external?

\code{df}: Degrees of freedom, i.e. the number of parameters.  Default 10.

\code{degree}: polynomial degree used for the basis function (default 3, a cubic)}

\item{modelid}{TODO}

\item{...}{Additional arguments to supply to control the Stan fit, passed to the \pkg{rstan} function \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}}
}
\description{
survextrap
}
