% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survextrap.R
\name{survextrap}
\alias{survextrap}
\title{survextrap}
\usage{
survextrap(
  formula,
  data,
  external = NULL,
  smooth_sd = "bayes",
  coefs_mean = NULL,
  cure = FALSE,
  cure_prior = c(1, 1),
  basehaz_ops = NULL,
  modelid = "mspline",
  fit_method = "mcmc",
  loo = (fit_method == "mcmc"),
  ...
)
}
\arguments{
\item{formula}{A survival formula in standard R formula syntax.  Covariates included
on the right hand side of the formula with be modelled with proportional hazards.}

\item{data}{Data frame containing variables in \code{formula}.}

\item{external}{External data as a data frame of aggregate survival counts with columns:

\code{start}: Start time

\code{start}: Follow-up time

\code{n}: Number of people alive at \code{start}

\code{r}: Number of those people who are still alive at \code{stop}

It is intended to add facilities to the package to produce this format of data from
other common forms of external data, e.g. registry or population data, or elicited judgements
about conditional survival}

\item{smooth_sd}{Smoothing parameter estimation.

\code{"bayes"}: the smoothing parameter is estimated by full Bayes.

\code{"eb"}: empirical Bayes is used.

Alternatively, if a number is supplied here, then the smoothing parameter is fixed to this number.}

\item{coefs_mean}{Spline basis coefficients that define the prior mean for the hazard function. By
default, these are set to values that define a constant hazard function.}

\item{cure}{If \code{TRUE}, a mixture cure model is used.}

\item{cure_prior}{Beta shape parameters for the cure probability (vector of two).}

\item{basehaz_ops}{A list of control parameters defining the spline model.

\code{iknots}: Internal knots.  If this is not supplied, then the number of knots is taken from \code{df},
and their location is taken from equally-spaced quantiles of the observed event times (concatenated
with the distinct follow-up times in the external data).

\code{bknots}: Boundary knots.  If this is not supplied, the boundary knots are set to zero and the
maximum event time (including the follow-up times in the external data).

\code{df}: Degrees of freedom, i.e. the number of parameters (or basis terms) that define the hazard
as a spline function of time.  Defaults to 10.  Note this does not necessarily overfit because
the function is smoothed through the prior.

\code{degree}: Polynomial degree used for the basis function. The default is 3, giving a cubic.}

\item{modelid}{\code{"mspline"} for the default M-spline model.

Only current alternative is \code{"weibull"} for a Weibull accelerated failure time model.
Just included for the purpose of package testing. Not fully tested, and not recommended for use in
practice for survival extrapolation.}

\item{fit_method}{Method from \pkg{rstan} used to fit the model.  Defaults to MCMC.

If \code{fit_method="mcmc"} then a sample from the posterior is drawn using Markov Chain Monte Carlo
sampling, via \pkg{rstan}'s \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}} function.
This is the default.  It is the most accurate but the slowest method.

If \code{fit_method="opt"}, then instead of an MCMC sample from the posterior,
\code{disbayes} returns the posterior mode calculated using optimisation, via
\pkg{rstan}'s \code{\link[rstan:stanmodel-method-optimizing]{rstan::optimizing()}} function.
A sample from a normal approximation to the (real-line-transformed)
posterior distribution is drawn in order to obtain credible intervals.

If \code{fit_method="vb"}, then variational Bayes methods are used, via \pkg{rstan}'s
\code{\link[rstan:stanmodel-method-vb]{rstan::vb()}} function.  This is labelled as "experimental" by
\pkg{rstan}.  It might give a better approximation to the posterior
than \code{fit_method="opt"}, but has not been investigated much for \code{disbayes} models.}

\item{loo}{Compute leave-one-out cross-validation statistics.  This is done by default for MCMC fits,
and not for optimisation or VB fits. Set to \code{FALSE} to save a bit of run time.
If these statistics are computed, then they are returned in the \code{loo} component of the
object returned by \code{survextrap}.}

\item{...}{Additional arguments to supply to control the Stan fit, passed to the appropriate
\pkg{rstan} function.}
}
\description{
survextrap
}
