% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survextrap.R
\name{survextrap}
\alias{survextrap}
\title{Flexible Bayesian parametric survival models}
\usage{
survextrap(
  formula,
  data,
  external = NULL,
  smooth_sd = "bayes",
  coefs_mean = NULL,
  cure = FALSE,
  nonprop = FALSE,
  prior_loghaz = p_normal(0, 20),
  prior_loghr = NULL,
  prior_smooth = p_gamma(2, 1),
  prior_cure = p_beta(1, 1),
  prior_logor_cure = NULL,
  prior_sdnp = p_gamma(2, 1),
  backhaz = NULL,
  mspline = NULL,
  fit_method = "mcmc",
  loo = (fit_method == "mcmc"),
  ...
)
}
\arguments{
\item{formula}{A survival formula in standard R formula syntax, with a call to \code{Surv()}
on the left hand side.

Covariates included on the right hand side of the formula with be modelled with
proportional hazards.}

\item{data}{Data frame containing variables in \code{formula}.}

\item{external}{External data as a data frame of aggregate survival counts with columns:

\code{start}: Start time

\code{stop}: Follow-up time

\code{n}: Number of people alive at \code{start}

\code{r}: Number of those people who are still alive at \code{stop}

It is intended to add facilities to the package to produce this format of data from
other common forms of external data, e.g. registry or population data, or elicited judgements
about conditional survival}

\item{smooth_sd}{Smoothing parameter estimation.

\code{"bayes"}: the smoothing parameter is estimated by full Bayes.

\code{"eb"}: empirical Bayes is used.

Alternatively, if a number is supplied here, then the smoothing parameter is fixed to this number.}

\item{coefs_mean}{Spline basis coefficients that define the prior mean for the hazard function. By
default, these are set to values that define a constant hazard function.  They are normalised to
sum to 1 internally (if they do not already).}

\item{cure}{If \code{TRUE}, a mixture cure model is used, where the "uncured" survival is defined by the
M-spline model, and the cure probability is estimated.}

\item{nonprop}{If \code{TRUE} then a non-proportional hazards model is fitted.
This is achieved by modelling the spline basis weights in terms of the covariates.  See
the "methods" vignette for more details.   In models with multiple covariates, currently
there is no way to assume that some covariates have proportional hazards but others don't -
it is all or none.}

\item{prior_loghaz}{Prior for the baseline log rate parameter (\code{log(eta)}), after centering around the log
crude event rate in the data.
This should be a call to a prior constructor function, such as
\code{p_normal(0,1)} or \code{p_t(0,2,2)}.   Supported prior distribution families
are normal (parameters mean and SD) and t distributions (parameters
location, scale and degrees of freedom).  The default is a normal distribution with
mean 0 and standard deviation 20.

"Baseline" is defined
by the continuous covariates taking a value of zero and factor covariates taking their
reference level.  To use a different baseline, the data should be transformed
appropriately beforehand, so that a value of zero has a different meaning.

For continuous covariates, it helps for both computation and interpretation to define the value of zero to
denote a typical value in the data, e.g. the mean.}

\item{prior_loghr}{Priors for log hazard ratios.  This should be a call to
\code{p_normal()} or \code{p_t()}.  A list of calls can also be provided, to give
different priors to different coefficients, where the name
of each list component matches the name of the coefficient, e.g.
\code{list("age45-59" = p_normal(0,1), "age60+" = p_t(0,2,3))}

The default is \code{p_normal(0,2.5)} for all coefficients.}

\item{prior_smooth}{Gamma prior for the smoothing standard deviation, in
models where this is estimated.  This should be a call to \code{p_gamma()}.  The
default is \code{p_gamma(2,1)}.}

\item{prior_cure}{Prior for the baseline cure probability.  This should be a
call to \code{p_beta()}.  The default is a uniform prior, \code{p_beta(1,1)}.
Baseline is defined by the mean of continuous covariates and the reference
level of factor covariates.}

\item{prior_logor_cure}{Priors for log odds ratios on cure probabilities.
This should be a call to \code{p_normal()} or \code{p_t()}.  The default is
\code{p_normal(0,2.5)}.}

\item{prior_sdnp}{Prior for the standard deviation parameters that smooth the non-proportionality
effects over time in non-proportional hazards models.  This should be a call to \code{p_gamma()}
or a list of calls to \code{p_gamma()} with one component per covariate, as in \code{prior_loghr}.}

\item{backhaz}{Background hazard, that is, for causes of death other than the
cause of interest. This defines a "relative survival" model where the
overall hazard is the sum of a cause-specific hazard and a background
hazard.   The background hazard is assumed to be known, and the
cause-specific hazard is modelled with the flexible parametric model.

The background hazard can be supplied in two forms.  The meaning of predictions
from the model depends on which of these is used.

(a) A data frame with columns \code{"hazard"} and \code{"time"}, specifying
the "background" hazard at all times as a piecewise-constant (step) function.
Each row gives the "background" hazard between the specified time and the
next time.  The first element of \code{"time"} should be 0, and the final row
specifies the hazard at all times greater than the last element of
\code{"time"}.

Predictions from the modelled fitted by \code{survextrap} will then include this
background hazard, because it is known at all times.

(b) The (quoted) name of a variable in the data giving the background
hazard.  For censored cases, the exact value does
not matter.  The predictions from \code{survextrap} will
then describe the excess hazard or survival on top of this background.
The overall hazard cannot be predicted in general, because the background hazard is
only specified over a limited range of time.

If there is external data and \code{backhaz} is supplied in form (b), then the
user should also supply the background survival at the start and stop points
in columns of the external data named \code{"backsurv_start"} and
\code{"backsurv_stop"}.  This should describe the same reference population as
\code{backhaz}, though the package does not check for consistency between these.

Leave-one-out cross-validation currently does not take into account any
background hazard (effectively assuming it to be zero).}

\item{mspline}{A list of control parameters defining the spline model.

\code{iknots}: Internal knots.  If this is not supplied, then the number of
knots is taken from \code{df}, and their location is taken from equally-spaced
quantiles of the observed event times (concatenated with the distinct
follow-up times in the external data).

\code{bknots}: Boundary knots.  If this is not supplied, the boundary knots are
set to zero and the maximum event time (including the follow-up times in
the external data).

\code{df}: Degrees of freedom, i.e. the number of parameters (or basis terms)
that define the hazard as a spline function of time.  Defaults to 10.  Note
this does not necessarily overfit because the function is smoothed through
the prior.

\code{degree}: Polynomial degree used for the basis function. The default is 3, giving a cubic.}

\item{fit_method}{Method from \pkg{rstan} used to fit the model.  Defaults to MCMC.

If \code{fit_method="mcmc"} then a sample from the posterior is drawn using Markov Chain Monte Carlo
sampling, via \pkg{rstan}'s \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}} function.
This is the default.  It is the most accurate but the slowest method.

If \code{fit_method="opt"}, then instead of an MCMC sample from the posterior,
\code{survextrap} returns the posterior mode calculated using optimisation, via
\pkg{rstan}'s \code{\link[rstan:stanmodel-method-optimizing]{rstan::optimizing()}} function.
A sample from a normal approximation to the (real-line-transformed)
posterior distribution is drawn in order to obtain credible intervals.

If \code{fit_method="vb"}, then variational Bayes methods are used, via \pkg{rstan}'s
\code{\link[rstan:stanmodel-method-vb]{rstan::vb()}} function.  This is labelled as "experimental" by
\pkg{rstan}.  It might give a better approximation to the posterior
than \code{fit_method="opt"}, but has not been investigated in depth for these models.}

\item{loo}{Compute leave-one-out cross-validation statistics.  This is done by default. Set to
\code{FALSE} to not compute them.
If these statistics are computed, then they are returned in the \code{loo} component of the
object returned by \code{survextrap}.  See the \code{"examples"} vignette for some explanation
of these.}

\item{...}{Additional arguments to supply to control the Stan fit, passed to the appropriate
\pkg{rstan} function, depending on which is chosen through the \code{fit_method} argument.}
}
\value{
A list of objects defining the fitted model.  A particularly important component is

\code{stanfit}: The object returned by \code{rstan} containing samples from the fitted model.  See the \href{https://mc-stan.org/rstan/index.html}{rstan} documentation.
}
\description{
Flexible Bayesian parametric survival models.  Individual data are represented using M-splines
and a proportional hazards or flexible non-proportional hazards model.
Optionally a mixture cure version
of this model can be fitted.
Extrapolations can be enhanced by including external aggregate data, or by including a fixed background hazard
in an additive hazards (relative survival) model.
}
