% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspline_dfns.R
\name{Survmspline}
\alias{Survmspline}
\alias{dsurvmspline}
\alias{psurvmspline}
\alias{qsurvmspline}
\alias{rsurvmspline}
\alias{hsurvmspline}
\alias{Hsurvmspline}
\alias{mean_survmspline}
\alias{rmst_survmspline}
\title{M-spline survival distribution}
\usage{
psurvmspline(
  q,
  alpha,
  coefs,
  knots,
  degree = 3,
  lower.tail = TRUE,
  log.p = FALSE
)

Hsurvmspline(x, alpha, coefs, knots, degree = 3, log = FALSE)

hsurvmspline(x, alpha, coefs, knots, degree = 3, log = FALSE)

dsurvmspline(x, alpha, coefs, knots, degree = 3, log = FALSE)

qsurvmspline(
  p,
  alpha,
  coefs,
  knots,
  degree = 3,
  lower.tail = TRUE,
  log.p = FALSE
)

rsurvmspline(n, alpha, coefs, knots, degree = 3)

rmst_survmspline(t, alpha, coefs, knots, degree = 3)

mean_survmspline(alpha, coefs, knots, degree = 3)
}
\arguments{
\item{alpha}{Log scale parameter.}

\item{coefs}{Spline basis coefficients. These should sum to 1,
otherwise they are normalised internally to sum to 1.}

\item{knots}{Locations of knots on the axis of time, supplied in
increasing order.  These include the two boundary knots.

In vectorised usage of these functions, the knots and degree must be
the same for all alternative times and parameter values.}

\item{degree}{Spline polynomial degree.}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P(X
\le x)}{P(X <= x)}, otherwise, \eqn{P(X > x)}{P(X > x)}.}

\item{x, q, t}{Vector of times.}

\item{log, log.p}{Return log density or probability.}

\item{p}{Vector of probabilities.}

\item{n}{Number of random numbers to simulate.}
}
\value{
\code{dsurvmspline} gives the density, \code{psurvmspline} gives the
distribution function, \code{hsurvmspline} gives the hazard and
\code{Hsurvmspline} gives the cumulative hazard.

\code{qsurvmspline} gives the quantile function, which is computed by
numerical inversion.

\code{rsurvmspline} generates random survival times by using
\code{qsurvmspline} on a sample of uniform random numbers.
}
\description{
Probability density, distribution, quantile, random generation, hazard,
cumulative hazard, mean and restricted mean functions for the
M-spline time-to-event model.
}
\details{
These are the same as the M-splines used to model survival data in \code{rstanarm}, except that an
additional assumption is made that the hazard is constant beyond the boundary knots at its
value at the boundary.   This gives a continuous but non-smooth function.
}
\references{
Ramsay, J. O. (1988). Monotone regression splines in action. Statistical Science, 3(4): 425–441.

Brilleman, S. L., Elci, E. M., Novik, J. B., & Wolfe, R. (2020). Bayesian survival analysis using the rstanarm R package. arXiv preprint arXiv:2002.09633.

Wang, W., Yan, J. (2021). Shape-restricted regression splines with R package splines2. Journal of Data Science_, \emph{19}(3), 498-517. doi:10.6339/21-JDS1020 \url{https://doi.org/10.6339/21-JDS1020}.
}
\author{
Christopher Jackson \href{mailto:chris.jackson@mrc-bsu.cam.ac.uk}{chris.jackson@mrc-bsu.cam.ac.uk}
}
\keyword{distribution}
